Fix SQLAlchemy Textual SQL expression error
Flask-SQLAlchemy does not version lock SQLAlchemy, and SQLAlchemy 1.3.x
was recently released which introduced that error. Prior to 1.3.x it was
output as a warning instead of an error.

This change hard locks SQLAlchemy to 1.2.0 and a future update video
will adjust the app's code so it works with SQLAlchemy 1.3.x and beyond.